Factor command implementation to separate package.
